+++ package.json
@@ -8,7 +8,7 @@
     "bundle": "bin/bundle.js",
     "pretest": "npm run bundle",
     "version": "npm run bundle && git add .",
-    "test": "mocha"
+    "test": "mocha --parallel"
   },
   "repository": {
     "type": "git",
